// src/app/models/model/product-type-group.model.ts
export class ProductTypeGroupModel {
    id: number;
    title: string;
    description: string;
    priceWeight: number | null;
    priceQuantity: number | null;
    priceStandard: number | null;
    note: string;
    code: string;
    priceJp: number | null;
    priceUs: number | null;
    priceKr: number | null;
    priceAu: number | null;
    isEdit: boolean;
}